v1.8.1：--step 取值规则由 'numeric' 改 'integer'，小数步数不再被截断执行
v1.8.0 用 ['numeric','min:1'] 校验 --step，实测放过 --step=1.5，(int) 再截成 1，
在应用库上真的回滚掉了一个批次（撤的不是用户说的那一步数）。改为 ['integer','min:1']：
filter_var 直接拒掉小数与科学计数法。测试补 3 行（migrate 1.5 / 1e3、rollback 2.5），
mutant 脚本补 2 个（把规则改回 numeric 即红），README 与框架文档同步口径。

+++ src/Console/Commands/MigrateCommand.php
@@ -36,10 +36,10 @@ final class MigrateCommand extends Command
         $step = null;
         if ($this->input->provided('step')) {
             // 没写 --step 才是「不限步数」；写了就必须是个 ≥1 的整数。
-            // 放进 opt() 的默认值兜底会反过来：`(int) 'abc'` 是 0（一步都不跑）、
-            // 光秃秃的 `--step` 是 null（不限步数），两者都跟用户写的意思相反，且照样退出 0。
+            // 靠强转兜底会反向：`(int) 'abc'` 是 0（一步都不跑），`--step=1.5` 过 'numeric'
+            // 再被截成 1 —— 跑掉的都不是用户说的那一步数。
             $given = $this->opt('step');
-            if (!$this->input->validate('step', $given, ['numeric', 'min:1'])) {
+            if (!$this->input->validate('step', $given, ['integer', 'min:1'])) {
                 $this->error('--step 需要 ≥1 的整数，收到: ' . var_export($given, true));
 
                 return 1;
